+++ src/components/Card/CardContainer/CardContainer.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useGetMessage } from "../../../data-access/useGetMessage";
-import CardBody from "./CardBody";
+import CardBody from "../CardBody/CardBody";
 export const CardContainer = () => {
   const { data } = useGetMessage();
   return (
